I, like an idiot published my bot token
I accidentally published my bot token to GitHub leading to someone who doesn't know what a DDOS is taking control of my bot and thus leading to the total destruction of my Discord server.

It really isn't a big deal, since there's a good chance you're reading this. I have been needing to remove a lot of the stuff from the server for a long time now. Thanks for the fresh start!