Update to 3.8 and general fixes
I changed my discord name to Sidpatchy, I changed all mentions of Rainverm38.

For some reason I was mistakenly still using user.nick instead of user.name. This caused users without nicknames to get labeled as none when mentioned. Interestingly, the version my server runs uses user.name already, it seems I fixed this previously but forgot to apply to to the previous version of RomeBot.

Next I updated the last update time, that had been neglected previously...

Finally, I stopped importing the entire time library and now just import sleep. This is simply because I had misunderstood the initial problem that caused me to import time and sleep separately.

No fixes were required  specifically for 3.8 except for the time library, which was a simple fix, that perhaps, was supposed to be or already was a problem in previous versions of RomeBot.

A final note, because of the lack of breaks, I have verified that RomeBot will still run on 3.7, though I will no longer be actively supporting it. In the future if you encounter any issues running RomeBot on 3.7, please try 3.8, otherwise, I would be more than happy to help.

Actually finally, Interestingly, RomeBot seems to run a couple milliseconds faster on 3.8 than it did on 3.7. Not that it matters to anyone, but it's kinda interesting.