RomeBot v1.2.0
I finally got around to making a logging fuction. I run these bots in a screen session, it doesn't allow me to scoll up, unfortunately. Plus, this is more permanet and gave me something to do while I am stuck at home :(

This doesn't follow the format of the console outputs exactly, mainly because I think it looks cooler the way I have done it. The main difference is when the bot is started and when the bot is connected. Otherwise, everything is identical in look.

I also finally got around to changing the uptime command. For some reason I thought 2 lines looked cool but in reality it just gets spammy, espeically when you have all of my bots in the same server. The solution to this was pretty simple but it would have taken less time to figure out if I had bothered to read the docs. The same changes were made to the lastupdate command.a

I also questioned why chuga chuga choo choo was the sound the economy makes. Dunno, that's just what happened.

I changed the 'hitler = bad' to 'hitler was bad' because I feel like the = makes me sound like a neo-nazi, which, for the record, I am not. I doubt anyone actually thought that but in any case, I switched it. Cancel that. I decided to remove the clarification entirely. I feel like the whole thing made me look like a neo-nazi. Again, I doubt anyone thought I was.