This was a bad idea.
I don't know how to SQL. What I do know is limited. I probably should have used something else to achieve this. I don't care though. I am tired and want to go to bed so I'll keep this short.

RomeBot.py has grown by maybe 200 lines. This will probably shrink back down, though, nothing is more permanent than a temporary solution that works. Most of the changes are just what I have achieved so far in setting up a database. I haven't done any extensive testing on RomeBot with this but it actually seems relatively stable. If you are looking at this branch hoping to be able to use it, that's fine but do know that it is super alpha and is probably going to break on your system at some point. For now, just stick with the release version of RomeBot if you are doing anything other than testing. The compiler will not work currently because I simply don't feel like adding a # to the last line. If you want to compile it, for some lame reason, feel free to add that '#' and have a brilliant day, otherwise just have a look at the wiki, those guides still apply.

There have been two temporary files created. First, readDB.py. All it does is read RomeBot.db and print it in the console so that I don't need write out the code for that in RomeBot everytime I want to see what is in the database. Next, we have RomeBot COmmand Changes. This simply is a listing of new and changed commands, it currently only features the !registerServer command because it is the only one that is close to complete and isn't written down elsewhere, yet. More stuff will be added to this as I progress through this large project.

I have decided that this will be titled RomeBot v2.0.0 because of the absolutely massive amount of changes that are going to be required. Unless I go full #nolife for the next couple days, this probably won't be out for at least a week, probably two. There is a lot of work that needs to be done. I haven't even begun working on the features that I believe should be allowed to be turned off on certain servers.

Fuck me, I don't know why I thought this was a good idea.

Additionally, I have noticed a bug in the logging function, whenever the bot reconnects to discord it runs the calculation of how long it took to connect from when the bot was started. I will probably fix this in the release of v2.0.0 instead of having to worry about fixing up more than one branch at a time.

That wasn't at all breif.