Fixed database writting.
I'm gonna be comepletely honest with you here, I don't actually remember what was wrong but that's what my problem was last night and it works now so that's cool. That's only partially a lie.

The problem seemed to be my code writting to the database. That wasn't the case. It WAS writing to the database, it wasn't reading the database correctly though. I forgot to run a lookup in the database which was the issue. Everything after that was pretty simple.

I have changed my attack vector (aka how I want to achieve this) and I have gone for the route that involes the least amount of fiddling with SQLite which should actually be fastest. Currently, since there are 22 toggleable features, RomeBot has a string of 22  1s added to the database. In future, RomeBot will allow for updating the database, this will probably be automatic but I have created a placeholder command for the time being called !updateServer, this may end up remaining to allow the user to manually update their listing in the database.

Due to my switch in how I want to achieve feature control, I have removed almost everything I did last night. While this wasn't fun to do, it was required to ensure I don't just hack together a solution.

As always, I am trying to make this as easy as possible to host on your own. As a result, everything currently configures on its own, I hope this will continue to be the way things are handled.