I'm sane!
I have FINALLY perfected the function that not only handles toggling but also handles writing to the database and the other function that handles reading the database.  This needs to be implemented elsewhere, notably, in the readDB function or in its own fucntion entirely, this will allow for easy testing of whether or not a command is enabled or not. Scrap that, I may be wrong there. I think what I have done in readDB()[0] is reporting if the command is enabled or not, I am not entirely sure why I am not certain of this but I  will need to run some tests to be sure. This will unfortunately have to take place tommorow as I really don't feel like doing this anymore tonight and I've already worked on this for 7ish hours today and worked at work for another 4ish hours, I'm pooped and will probably only cause this to take longer than it needs to if I keep working tonight. By the looks of things right now, I may have this out within a week although, I probably just jynxed it.

Something else I would like to add is the ability to enable or disable RomeBot in  certain chat channels this would require either another table in the database or something entirely different.  That might be a project for another update though, it may never even see the light of day since you can easily do that in Discord already.

Goodnight!

Sorry, one more thing, I am trying something new with GitHub desktop, I have clicked on the lines that include the changes to the bot.run command, I don't think it will commit them, but either way, if it does, I will just generate a new token and move on, also, it's a token for my TestBot and is only used in my testing servers.