v2-a1
The first "release" of RomeBot v2. I removed most of the commands and added comments after each line of the help command indicating my intentions for the future of each command. They aren't set in stone yet but are likely close-ish to what will happen. Unusable commands are commented out, this shouldn't have any affect on anyone, as no one should use this version... it only contains !help and !version both of which will likely be changed.