v2-a2
Implemeted sLOUT in places that didn't get replaced right away. Added a feature to sLOUT that reads the configuration file... I might have to make this a slightly modified version of sLOUT just so I am not reading every file that ends in .bot in the directory. That's a temporary thing, I will come up with a better solution eventually. Additionally, I added an info command, made the version command work again, and reworded parts of !help. Most of the work in this update was does in sLOUT...

One more thing, romebot.bot is very picky about formatting right now. If you remove the space before or after the = sign in the config file, it will break. There will be a fix for this in the next version.