Add most commands back
While neglecting the whole reason I decided to rewrite this, I added most commands back to the bot. The ones that are available should be in the help command (that honestly still needs to be looked through).

I've updated to sLOUT v0.1-a10 (which is honestly just v0.1 at this point), this means that there are some differences in RomeBot that weren't there the last time I commited to this repo.  One of the major ones is switching from using my own method to read from a file to a proper method (yaml, in this case). This means various things that I don't feel like typing out, go read the sLOUT update log.

I've also decided to let RomeBot send images as files instead of linking from Imgur as doing so allows me to more "nicely" send one message rather than multiple.