v2-RC1 | final touches
Some of the changes here took place a few months ago, and a most were implemented tonight.

Most of the changes include neatening commands that were previously glorified copy-pastes from the old version of RomeBot.

Removed a handful of commands.

Made the info command look glorious compared to that old piece of crap (https://i.imgur.com/xZoa8bp.png)

and probably a lot of other stuff.

The next commit to this branch will likely be the final, after that everything will be merged into the master branch.