Start work on a polling function
This is only the basic framework of what I plan on having in this command. It *technically* works as is, however, I'd like to add a bit more to it. A section will be added to the help command that discusses how to use the polling function more in depth. That's all, I think.